merge-lock-status-service/requirements.txt
python-dotenv==0.6.1
